test-qenetex-api/pyproject.toml
[tool.poetry]
name = "test-api"
version = "0.1.0"
description = "test api"
authors = ["Andrey Smirnov <spiehdid@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
SQLAlchemy = "^1.4.3"
psycopg2 = "^2.8.6"
alembic = "^1.5.8"
FastAPI-SQLAlchemy = "^0.2.1"
python-dotenv = "^0.16.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
